feat: HistoryStorage 타입 생성 및 구현
- 기존 MemoStorage에 있던 History 가져오기 기능 HistoryStorage로 이관
- HistoryRepository와 HistoryStorage를 연결(HistoryStorageable을 이용한 의존성 역전)
- MemoStorageable 프로토콜 파일의 위치 변경
